# PE: Getting exports speedup