--- a/libr/bin/format/pe/pe.c
+++ b/libr/bin/format/pe/pe.c
@@ -3096,21 +3096,28 @@ struct r_bin_pe_export_t* PE_(r_bin_pe_get_exports)(struct PE_(r_bin_pe_obj_t)* 
 		functions_paddr = bin_pe_rva_to_paddr (bin, bin->export_directory->AddressOfFunctions);
 		names_paddr = bin_pe_rva_to_paddr (bin, bin->export_directory->AddressOfNames);
 		ordinals_paddr = bin_pe_rva_to_paddr (bin, bin->export_directory->AddressOfOrdinals);
+
+		const size_t names_sz = bin->export_directory->NumberOfNames * sizeof (PE_Word);
+		const size_t funcs_sz = bin->export_directory->NumberOfFunctions * sizeof (PE_VWord);
+		PE_Word *ordinals = malloc (names_sz);
+		PE_VWord *func_rvas = malloc (funcs_sz);
+		if (!ordinals || !func_rvas) {
+			free (exports);
+			free (ordinals);
+			free (func_rvas);
+			return NULL;
+		}
+		r_buf_read_at (bin->b, ordinals_paddr, (ut8 *)ordinals, names_sz);
+		r_buf_read_at (bin->b, functions_paddr, (ut8 *)func_rvas, funcs_sz);
 		for (i = 0; i < bin->export_directory->NumberOfFunctions; i++) {
 			// get vaddr from AddressOfFunctions array
-			int ret = r_buf_read_at (bin->b, functions_paddr + i * sizeof(PE_VWord), (ut8*) &function_rva, sizeof(PE_VWord));
-			if (ret < 1) {
-				break;
-			}
+			function_rva = r_read_at_ble32 ((ut8 *)func_rvas, i * sizeof (PE_VWord), bin->endian);
 			// have exports by name?
 			if (bin->export_directory->NumberOfNames != 0) {
 				// search for value of i into AddressOfOrdinals
 				name_vaddr = 0;
 				for (n = 0; n < bin->export_directory->NumberOfNames; n++) {
-					ret = r_buf_read_at (bin->b, ordinals_paddr + n * sizeof(PE_Word), (ut8*) &function_ordinal, sizeof (PE_Word));
-					if (ret < 1) {
-						break;
-					}
+					function_ordinal = r_read_at_ble16 ((ut8 *)ordinals, n * sizeof (PE_Word), bin->endian);
 					// if exist this index into AddressOfOrdinals
 					if (i == function_ordinal) {
 						// get the VA of export name  from AddressOfNames
@@ -3158,6 +3165,8 @@ struct r_bin_pe_export_t* PE_(r_bin_pe_get_exports)(struct PE_(r_bin_pe_obj_t)* 
 			exports[i].last = 0;
 		}
 		exports[i].last = 1;
+		free (ordinals);
+		free (func_rvas);
 	}
 	exp = parse_symbol_table (bin, exports, exports_sz - sizeof (struct r_bin_pe_export_t));
 	if (exp) {