<!-- Filling this template is mandatory -->

**Your checklist for this pull request**
- [x] I've read the [guidelines for contributing](https://github.com/radareorg/radare2/blob/master/DEVELOPERS.md) to this repository
- [x] I made sure to follow the project's [coding style](https://github.com/radareorg/radare2/blob/master/DEVELOPERS.md#code-style)
- [ ] I've added tests that prove my fix is effective or that my feature works (if possible)
- [ ] I've updated the documentation and the [radare2 book](https://github.com/radareorg/radare2book) with the relevant information (if needed)

**Detailed description**

<!-- Explain the **details** for making this change. Is a new feature implemented? What existing problem does the pull request solve? How does the pull request solve these issues? Please provide enough information so that others can review your pull request. -->

## Before
```
radare2 -qq /mnt/c/Windows/System32/ucrtbased.dll 
1.11s user 0.03s system 98% cpu 1.153 total
```
## After
```
radare2 -qq /mnt/c/Windows/System32/ucrtbased.dll 
0.39s user 0.02s system 96% cpu 0.420 total
```

**Test plan**

Tests pass