@radare 2020-05-03
Ideallyi the endian is the same as the host just keeping the way you were doing would be the most performant option. also not sure how many big endian PEs would be out there